ripping out core commands from general grammar test
core commands make a separate test file
cleaning up
adding target related command grammar tests
