+++ js/ohm/tests/test-grammar.js
@@ -304,117 +304,6 @@ describe("SimpleTalk Grammar", () => {
             semanticMatchFailTest("pa ram1, param2", "ParameterList");
         });
     });
-    describe("Commands", () => {
-        describe("Go To", () => {
-            it ("go to direction with no object fails", () => {
-                const direction = ["next", "previous"];
-                direction.forEach((d) => {
-                    const s = `go to ${d}`;
-                    semanticMatchFailTest(s, "Command");
-                });
-            });
-            it ("go to with object", () => {
-                const direction = ["next", "previous"];
-                direction.forEach((d) => {
-                    const s = `go to ${d} card`;
-                    semanticMatchTest(s, "Command");
-                    semanticMatchTest(s, "Command_goToDirection");
-                    semanticMatchTest(s, "Statement");
-                });
-            });
-            it ("go to by reference", () => {
-                const s = "go to card 42";
-                semanticMatchTest(s, "Command");
-                semanticMatchTest(s, "Command_goToByReference");
-                semanticMatchTest(s, "Statement");
-            });
-            it.skip ("Bad go to: invalid object", () => {
-                const s = "go to world 42";
-                semanticMatchFailTest(s, "Command");
-            });
-            it ("Bad go to: invalid structure", () => {
-                const s = "go to next card 42";
-                semanticMatchFailTest(s, "Command");
-            });
-        });
-        describe("Remove Model", () => {
-            it ("Basic Remove (no id)", () => {
-                const direction = ["stack", "background", "card", "button"];
-                direction.forEach((d) => {
-                    const s = `delete this ${d}`;
-                    semanticMatchTest(s, "Command");
-                    semanticMatchTest(s, "Command_deleteModel");
-                    semanticMatchTest(s, "Statement");
-                });
-            });
-            it ("Basic Remove (with id)", () => {
-                const direction = ["stack", "background", "card", "button"];
-                direction.forEach((d) => {
-                    const s = `delete ${d} 20`;
-                    semanticMatchTest(s, "Command");
-                    semanticMatchTest(s, "Command_deleteModel");
-                    semanticMatchTest(s, "Statement");
-                });
-            });
-            it.skip ("Bad delete (world)", () => {
-                const s = "delete this world";
-                semanticMatchFailTest(s, "Command_deleteModel");
-                semanticMatchFailTest(s, "Command");
-            });
-        });
-        describe("Answer", () => {
-            it ("simple answer", () => {
-                const s = "answer \"42\"";
-                semanticMatchTest(s, "Command");
-                semanticMatchTest(s, "Command_answer");
-                semanticMatchTest(s, "Statement");
-            });
-            it ("bad answer", () => {
-                const s = "answer \"42\" \"42\"";
-                semanticMatchFailTest(s, "Command");
-            });
-        });
-        describe("Set", () => {
-            it.skip ("Set someProperty with id", () => {
-
-                objects.forEach((d) => {
-                    const s = `set "someProperty" to "blue" in ${d} 10`;
-                    semanticMatchTest(s, "Command");
-                    semanticMatchTest(s, "Command_setProperty");
-                    semanticMatchTest(s, "Statement");
-                });
-            });
-            it ("Set someProperty (no target)", () => {
-                objects.forEach((d) => {
-                    const s = `set "someProperty" to "blue"`;
-                    semanticMatchTest(s, "Command");
-                    semanticMatchTest(s, "Command_setProperty");
-                    semanticMatchTest(s, "Statement");
-                });
-            });
-            it.skip("Set background color in this or current", () => {
-                ["this", "current"].forEach((context) => {
-                    objects.forEach((d) => {
-                        const s = `set "someProperty" to "blue" in ${context} ${d}`;
-                        semanticMatchTest(s, "Command");
-                        semanticMatchTest(s, "Command_setProperty");
-                        semanticMatchTest(s, "Statement");
-                    });
-                });
-            });
-        });
-        describe("Arbitrary", () => {
-            it ("arbitrary command", () => {
-                const s = "anythinggoes";
-                semanticMatchTest(s, "Command");
-                semanticMatchTest(s, "Statement");
-            });
-        });
-        it ("Bad command (arbitrary with digits)", () => {
-            semanticMatchFailTest("1234arrowKe", "Command");
-        });
-    });
-
     describe("Variable assignment", () => {
         it("Can parse put assignment command (string literal)", () => {
             const s = `put "hello" into myCustomVariable`;
